% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spk_join.R
\name{spk_join}
\alias{spk_join}
\title{Spatial Join with Optional Mask Filtering and Column Selection}
\usage{
spk_join(
  target_tbl,
  mask_tbl,
  target_col_return = "*",
  mask_col_return = NULL,
  mask_col_filter = NULL,
  mask_col_filter_values = NULL,
  mask_col_filter_values_negate = FALSE,
  join_fun = sf::st_intersects,
  path_gpkg = NULL,
  collapse = FALSE,
  target_col_collapse = NULL,
  ...
)
}
\arguments{
\item{target_tbl}{\link[sf:sf]{sf::sf} The target table to be spatially joined.}

\item{mask_tbl}{\link{character} or \link[sf:sf]{sf::sf} The mask table name (if reading from \code{path_gpkg}) or an sf object.}

\item{target_col_return}{\link{character} A vector of target column names to retain. Use "*" to retain all.}

\item{mask_col_return}{\link{character} or \link{NULL} Optional. A vector of mask column names to retain. Default is \code{NULL}.}

\item{mask_col_filter}{\link{character} or \link{NULL} Optional. A single mask column name used for filtering rows.}

\item{mask_col_filter_values}{\link{character} or \link{NULL} Optional. Values to match (or exclude) in \code{mask_col_filter}.}

\item{mask_col_filter_values_negate}{\link{logical} Whether to exclude (\code{TRUE}) or include (\code{FALSE}) matching rows. Default is \code{FALSE}.}

\item{join_fun}{\link{function} Spatial predicate function from \code{\link[sf:geos_binary_pred]{sf::st_intersects()}}, \code{\link[sf:geos_binary_pred]{sf::st_within()}}, etc. Default is \code{sf::st_intersects}.}

\item{path_gpkg}{\link{character} or \link{NULL} Optional. Path to a GeoPackage to read \code{mask_tbl} from if not already an sf object.}

\item{collapse}{\link{logical} Whether to collapse results to one amalgamated row with multiple results when more than
one result is returned. Default is TRUE}

\item{target_col_collapse}{\link{character} or \link{NULL} Optional. A target column name to group by to facilitate collapsing the result.}

\item{...}{Additional arguments passed to \code{\link[sf:st_join]{sf::st_join()}}.}
}
\value{
A sf object with the result of the spatial join. If \code{target_col_return} is not "*", only selected columns are returned.
}
\description{
Joins a target \link[sf:sf]{sf::sf} object with a mask \link[sf:sf]{sf::sf} object using a specified spatial predicate.
Optional filtering and selection of columns from the mask can be applied.
}
\seealso{
Other spacehakr: 
\code{\link{spk_gdalwarp}()},
\code{\link{spk_layer_info}()},
\code{\link{spk_odm}()},
\code{\link{spk_poly_to_points}()},
\code{\link{spk_q_layer_info}()},
\code{\link{spk_rast_ext}()},
\code{\link{spk_rast_not_empty}()},
\code{\link{spk_rast_rm_empty}()},
\code{\link{spk_res}()},
\code{\link{spk_stac_calc}()}
}
\concept{spacehakr}
