% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spk_rast_not_empty.R
\name{spk_rast_not_empty}
\alias{spk_rast_not_empty}
\title{Check if Raster Has Non-Zero Data}
\usage{
spk_rast_not_empty(f)
}
\arguments{
\item{f}{\link{character} A single file path to a raster file.}
}
\value{
\link{logical} \code{TRUE} if the raster has any non-zero values, otherwise \code{FALSE}.
}
\description{
Determines whether a raster file contains any non-zero values.
}
\seealso{
Other raster: 
\code{\link{spk_rast_ext}()},
\code{\link{spk_rast_rm_empty}()},
\code{\link{spk_res}()}
}
\concept{raster}
