% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spk_res.R
\name{spk_res}
\alias{spk_res}
\title{Extract Resolution from a Raster}
\usage{
spk_res(path, crs_out = "EPSG:3005")
}
\arguments{
\item{path}{\link{character} A file path to the input raster.}

\item{crs_out}{\link{character} The desired CRS (default is "EPSG:3005").}
}
\value{
\link{numeric} A vector of length two representing the resolution of the single pixel in metres in the specified CRS.
}
\description{
This function calculates and extracts the resolution (in metres) of a single pixel (centre) from a raster file.
It determines the central pixel of the raster, crops to that single pixel, and
returns its resolution in the specified CRS. Uses \code{\link[terra:rast]{terra::rast()}}, \code{\link[terra:ext]{terra::ext()}}, \code{\link[terra:crop]{terra::crop()}}, and \code{\link[terra:project]{terra::project()}}
to perform these operations.
}
\details{
The function first calculates the extent and resolution of the raster, identifies the central pixel,
and crops the raster to that pixel. It then projects the cropped raster to the specified CRS and
calculates the resolution of the single pixel.
}
\examples{
path <- system.file("extdata", "test1.tif", package = "spacehakr")
crs_out <- "EPSG:32609"
spk_res(path, crs_out)

}
\seealso{
Other raster: 
\code{\link{spk_rast_ext}()},
\code{\link{spk_rast_not_empty}()},
\code{\link{spk_rast_rm_empty}()}
}
\concept{raster}
